requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.14.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
daphne==4.1.0
defusedxml==0.8.0rc2
Django==4.2.8
django-money==3.4.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-yasg==1.21.7
Flask==3.0.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.1
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.3
msgpack==1.0.7
oauthlib==3.2.2
packaging==23.2
py-moneyed==3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-dotenv==1.0.0
python-ipware==2.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
runserver==1.0.4
service-identity==24.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.1
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==3.0.1
zipp==3.17.0
zope.interface==6.2
